Made text weight to actually render

+++ crates/core/src/padding.rs
@@ -1,13 +1,14 @@
-#[derive(Debug)]
+#[derive(Debug, Default, Clone)]
+/// Space inside an element
 pub struct Padding {
-    pub top: f32,
-    pub left: f32,
-    pub right: f32,
-    pub bottom: f32,
+    pub top: i32,
+    pub left: i32,
+    pub right: i32,
+    pub bottom: i32,
 }
 
 impl Padding {
-    pub fn all(padding: f32) -> Self {
+    pub fn all(padding: i32) -> Self {
         Padding {
             top: padding,
             left: padding,
@@ -18,29 +19,29 @@ impl Padding {
 
     pub fn new() -> Self {
         Padding {
-            top: 0.0,
-            left: 0.0,
-            right: 0.0,
-            bottom: 0.0,
+            top: 0,
+            left: 0,
+            right: 0,
+            bottom: 0,
         }
     }
 
-    pub fn left(mut self, padding: f32) -> Self {
+    pub fn left(mut self, padding: i32) -> Self {
         self.left = padding;
         self
     }
 
-    pub fn right(mut self, padding: f32) -> Self {
+    pub fn right(mut self, padding: i32) -> Self {
         self.right = padding;
         self
     }
 
-    pub fn bottom(mut self, padding: f32) -> Self {
+    pub fn bottom(mut self, padding: i32) -> Self {
         self.bottom = padding;
         self
     }
 
-    pub fn top(mut self, padding: f32) -> Self {
+    pub fn top(mut self, padding: i32) -> Self {
         self.bottom = padding;
         self
     }
